Innfer function to sample MC trees with binary covariates

Innfer function to sample MC trees with binary covariates

sample_mc_binary_cov_inner(x_pos, x_neg, B, root_node = NULL)

Arguments

x_pos
number of positive individuals in cluster

x_neg
number of negative individuals in cluster

B
number of trees to sample

root_node
if NULL, use the base model. If 1, ensure that the root node is positive. If 0, ensure that the root node is negative.

Value

a data frame with the following columns

freq

frequency of the observed data

cluster_size

total size of the cluster

x_pos

number of individuals in the cluster with the feature of interest =1

x_neg

number of individuals in the cluster with the feature of interest =0

mc_freq

number of MC frequency of clusters

x_pos_trans

number of transmissions from a positive

x_neg_trans

number of transmissions from a negative

root_node

only for the multiple outsider. Is 1 if the outsider is positive and 0 if it is negative