Simulate the process of flipping until failure for K clusters

Simulate the process of flipping until failure for K clusters

simulate_general_outbreak(
  K,
  inf_params,
  sample_covariates_df,
  covariate_names,
  covariate_weights = NULL,
  max_size = 50
)

Arguments

K
number of total clusters to simulate

inf_params
vector with beta coefficients to use in logistic function for probability of transmission

sample_covariates_df
Data frame of covariates to sample from

covariate_names
names of the covariates. Must match size of inf_params - 1.

covariate_weights
default is NULL which draws uniformly at random with replacement from the sample_covariates_df. Otherwise, the weights are used.

max_size
maximum size a cluster can be

Value

data frame with the following columns

cluster_id

unique cluster ID

person_id

order of infection in the cluster

gen

generation number (>=0)

inf_id

ID of the infector

n_inf

number of people infected by person

censored

whether the cluster end was censored or not

cluster_size

size of the cluster

covariates

covariates of the individuals

Details

breadth not depth. Generate generation by generation as opposed to going up the branch til termination.

Examples

inf_params <- c("beta_0" = -2, "beta_1" = 1)