Subset of Tuberculosis data

See references for more details. The data is collected between 2003-2009 in Maryland, USA. This is originally from (Xie et al. 2018)

tb_clean

Format
<<<<<<< HEAD
The object is a data frame of dimension 1137x7. Each row is a detected case of TB.

sex

Sex of individual (F/M/Unknown)

county

County in Maryland of individual

race

Race of individual (Asian/Black or African American/White)

spsmear

Smear status of individual (positive/negative/unknown)

hivstatus

HIV status of individual (positive/negative/unknown)

homeless

Homelessness status of individual (yes/no)

groupr

Cluster group

rel_time

detection time relative to first individual in cluster

=======
The object is a data frame of dimension 1137x10. Each row is a detected case of TB.

sex

Sex of individual (F/M/Unknown)

county

county in MD

ageatrept

age of individual

race

race of individual (Asian/Black or African American/White)

spsmear

Smear status of individual (positive/negative/unknown)

hivstatus

HIV status of individual (positive/negative/unknown)

homeless

Homelessness status of individual (yes/no)

INIT_REGIMEN_START_DATE

date of start of treatment

PCR.Cluster

Cluster group (if any)

datecoll

Date sputum was collected

>>>>>>> fd400a1555a7a782265dd95dc3e8cb9f965adce0
References

Xie YL, Cronin WA, Proschan M, Oatis R, Cohn S, Curry SR, Golub JE, Barry III CE, Dorman SE (2018). “Transmission of Mycobacterium tuberculosis From Patients Who Are Nucleic Acid Amplification Test Negative.” Clinical Infectious Diseases, 67(11), 1653--1659.

Examples
<<<<<<< HEAD

head(tb_clean)

#> # A tibble: 6 x 8
#> # Groups:   group [1]
#>   sex    county       race            spsmear  hivstatus homeless group rel_time
#>   <chr>  <chr>        <chr>           <chr>    <fct>     <chr>    <chr>    <dbl>
#> 1 Male   "BALTIMORE … Black or Afric… "Negati… pos       Yes      2          375
#> 2 Male   "BALTIMORE … Black or Afric… "Negati… pos       Yes      2         1085
#> 3 Male   ""           Black or Afric… ""       unk       Yes      2         1357
#> 4 Male   "BALTIMORE … Black or Afric… "Positi… pos       Yes      2         1002
#> 5 Female "BALTIMORE … Black or Afric… "Negati… neg       No       2         1385
#> 6 Male   "BALTIMORE … Black or Afric… "Negati… pos       Yes      2         1029

summary(tb_clean)

#>      sex               county              race             spsmear         
#>  Length:389         Length:389         Length:389         Length:389        
=======
    

head(tb_clean)

#>      sex         county ageatrept                      race  spsmear hivstatus
#> 1   Male BALTIMORE CITY        35 Black or African American Negative  Positive
#> 2   Male BALTIMORE CITY        47 Black or African American Negative  Positive
#> 3   Male                          Black or African American            Unknown
#> 4   Male BALTIMORE CITY        50 Black or African American Positive  Positive
#> 5 Female BALTIMORE CITY        21 Black or African American Negative  Negative
#> 6   Male BALTIMORE CITY        50 Black or African American Negative  Positive
#>   homeless INIT_REGIMEN_START_DATE PCR.Cluster   datecoll
#> 1      Yes              2005-05-12           2 2005-05-15
#> 2      Yes              2007-06-28           2 2007-04-25
#> 3      Yes                    <NA>           2 2008-01-22
#> 4      Yes              2007-02-12           2 2007-02-01
#> 5       No              2008-02-27           2 2008-02-19
#> 6      Yes              2007-03-17           2 2007-02-28

summary(tb_clean)

#>      sex               county           ageatrept             race          
#>  Length:1137        Length:1137        Length:1137        Length:1137       
>>>>>>> fd400a1555a7a782265dd95dc3e8cb9f965adce0
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
<<<<<<< HEAD
#>  hivstatus   homeless            group              rel_time     
#>  neg:272   Length:389         Length:389         Min.   :   0.0  
#>  unk: 61   Class :character   Class :character   1st Qu.:   0.0  
#>  pos: 56   Mode  :character   Mode  :character   Median : 192.0  
#>                                                  Mean   : 463.1  
#>                                                  3rd Qu.: 885.0  
#>                                                  Max.   :1938.0  

=======
#>                                                                             
#>    spsmear           hivstatus           homeless        
#>  Length:1137        Length:1137        Length:1137       
#>  Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character  
#>                                                          
#>                                                          
#>                                                          
#>                                                          
#>  INIT_REGIMEN_START_DATE PCR.Cluster           datecoll         
#>  Min.   :2003-12-19      Length:1137        Min.   :2003-12-22  
#>  1st Qu.:2005-05-03      Class :character   1st Qu.:2005-04-19  
#>  Median :2006-10-11      Mode  :character   Median :2006-10-12  
#>  Mean   :2006-10-09                         Mean   :2006-10-06  
#>  3rd Qu.:2008-03-18                         3rd Qu.:2008-03-12  
#>  Max.   :2009-09-29                         Max.   :2009-08-28  
#>  NA's   :363                                NA's   :357         

>>>>>>> fd400a1555a7a782265dd95dc3e8cb9f965adce0